Flytt felles datoutil ut til utilfil
